pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "trafilatura>=2.0.0",
]
